requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
django-filter==24.3
qrcode==7.4.2
reportlab==4.2.2
openpyxl==3.1.5
django-cors-headers==4.4.0
drf-yasg==1.21.7         
